Allow default option for smaller chromosome; update egg-game
We can still define chromosome image dimensions by hand, but we can also
pass in a `small=true` property, which will switch to a smaller set of
defaults.

We use this smaller version in the egg game.

Also made the non-editable allele labels smaller, and improved its layout,
as well as making the default large chromosomes slightly smaller.