Load images before entering rooms
When the Challenge Container loads, or when the room changes, we check
to see if there are images to load. If there are, we load them while
displaying a new LoadingView, before rendering the regular page.

The images are listed in a file, because

1. We can control which images are loaded for each room,
2. We can be smarter about which ones to load, and not load tiny
images that we don't care if they appear late. This allows us to
speed up loading at the expense of having all the assets ready.

However, the problem with a list is that it needs to be kept up-to-date.
We'll see if this remains a problem.